cos/flow_hash.rs: colocate 8 tests with production (#984 P3 Phase 3c)
Move 8 tests from tx/mod.rs::tests to cos/flow_hash.rs::tests covering
cos_flow_hash_seed_from_os and the exact_cos_flow_bucket distribution/
stability suite.

Adds COS_FLOW_FAIR_BUCKETS import.

Cleanups (in line with Codex round-1 + Copilot review):
- Drop stale cos_flow_hash_seed_from_os import from tx/mod.rs and from
  the cos/mod.rs cfg-test re-export (only consumer was the moved test).
- Update bucket-count comment in distribution-divergence test from
  '64-bucket / 1/64 collision rate' to '1024-bucket / 1/1024'.
- Rename cos_flow_hash_seed_from_os_draws_nonzero_entropy →
  cos_flow_hash_seed_from_os_never_returns_zero and tighten the
  assertion to match the API's actual contract (zero→one remap on
  every call), not the looser 'any-nonzero in 4 draws' floor.

cargo test --bins 865/0/2.
